Nest PseudoElement inside TripleaHandler
